fix bug isus

+++ src/src/Tower/Bullet.java
@@ -1,28 +1,28 @@
 package Tower;
 import java.awt.Color;
 import java.awt.Graphics2D;
-
 import Enemy.Enemy;
 
 public class Bullet {
-    private int x, y;
+    private double x, y; 
     private Enemy target;
-    private int speed = 10;
+    private double speed = 10.0; 
     public boolean isActive = true;
-
-    public Bullet(int x, int y, Enemy target) {
+    public Bullet(double x, double y, Enemy target) {
         this.x = x;
         this.y = y;
         this.target = target;
     }
+
     public void update() {
         if (target == null || target.isDead()) {
             isActive = false;
             return;
         }
-        double diffX = target.getX() - x;
+        double diffX = target.getX() - x; 
         double diffY = target.getY() - y;
         double distance = Math.sqrt(diffX*diffX + diffY*diffY);
+        
         if (distance <= speed) {
             target.takeDamage(10); 
             isActive = false; 
@@ -31,8 +31,9 @@ public class Bullet {
             y += (diffY / distance) * speed;
         }
     }
+
     public void draw(Graphics2D g2) {
         g2.setColor(Color.YELLOW);
-        g2.fillOval(x - 4, y - 4, 8, 8);
+        g2.fillOval((int)x - 4, (int)y - 4, 8, 8); 
     }
 }
